Refactor Data Source and add Cypher
